Rewrite tests for memory overcommit
