--- tests/integration/test_global_overcommit_tracker/configs/global_overcommit_tracker.xml
@@ -1,3 +1,0 @@
-<clickhouse>
-    <max_server_memory_usage>50000000</max_server_memory_usage>
-</clickhouse>
