+++ tests/integration/test_overcommit_tracker/configs/global_overcommit_tracker.xml
@@ -1,0 +1,3 @@
+<clickhouse>
+    <max_server_memory_usage>2000000000</max_server_memory_usage>
+</clickhouse>
