--- tests/queries/0_stateless/02104_overcommit_memory.reference
@@ -1,1 +1,0 @@
-OVERCOMMITED WITH USER LIMIT WAS KILLED
